+++ src/User/Form/RegisterForm.php
@@ -9,8 +9,6 @@ declare(strict_types = 1);
 
 namespace Frontend\User\Form;
 
-use Zend\Stdlib\ArrayUtils;
-
 /**
  * Class RegisterForm
  * @package App\User\Form
@@ -20,15 +18,26 @@ class RegisterForm extends \Dot\User\Form\RegisterForm
     public function init()
     {
         parent::init();
-        $validationGroup = $this->getValidationGroup();
-        $validationGroup = ArrayUtils::merge($validationGroup, [
+        // if using the form helper template, form element order will match the order in the validation group
+        $validationGroup = [
+            'register_csrf',
             'user' => [
+                'username',
+                'email',
                 'details' => [
                     'firstName',
                     'lastName',
-                ]
+                ],
+                'password',
+                'passwordConfirm',
             ]
-        ]);
+        ];
+        if ($this->has('captcha')) {
+            $validationGroup[] = 'captcha';
+        }
+        // add submit to validation group,
+        // not needed usually bu needed for the form display helper partial template
+        $validationGroup[] = 'submit';
 
         $this->setValidationGroup($validationGroup);
     }