+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.1.1 - 2017-03-12
+
+Updated account and register forms to use the new form template helpers
+
+### Added
+* Form partial template helpers
+
+### Deprecated
+* Nothing
+
+### Removed
+* Nothing
+
+### Fixed
+* Nothing
+
+
 ## 0.1.0 - 2017-03-11
 
 Initial tagged release